Project 14. Checking on multiplicity

+++ Project3_MaxOfTwo/Program.cs
@@ -1,10 +1,10 @@
 ﻿// С клавиатуры вводятся два числа a и b. Найти максимальное из них. 
 Console.WriteLine("Привет! Это ПО ищет наибольшее число из двух.");
 Console.Write("Введите первое число: ");
-int a, b, max;
-a = Convert.ToInt32(Console.ReadLine());
+double a, b, max;
+a = Convert.ToDouble(Console.ReadLine());
 Console.Write("Введите второе число: ");
-b = Convert.ToInt32(Console.ReadLine());
+b = Convert.ToDouble(Console.ReadLine());
 if (a > b)
 {
   max = a;
